third_party/blink/web_tests/external/wpt/digital-credentials/create.tentative.https-expected.txt

This is a testharness.js-based test.
[FAIL] navigator.credentials.create() API rejects if there are no credential request.
  promise_rejects_js: function "function() { throw e; }" threw object "NetworkError: Error retrieving a token." ("NetworkError") expected instance of function "function TypeError() { [native code] }" ("TypeError")
[FAIL] navigator.credentials.create() API rejects if there are no credential request for same-origin iframe.
  promise_rejects_js: function "function() { throw e; }" threw object "NetworkError: Error retrieving a token." ("NetworkError") expected instance of function "function TypeError() { [native code] }" ("TypeError")
[FAIL] navigator.credentials.create() API rejects if there are no credential request in cross-origin iframe.
  assert_equals: expected "TypeError" but got "DOMException"
[FAIL] navigator.credentials.create() promise is rejected if called with an aborted controller.
  promise_rejects_dom: function "function() { throw e; }" threw object "NotSupportedError: Only exactly one of 'password', 'federated', and 'publicKey' credential types are currently supported." that is not a DOMException AbortError: property "code" is equal to 9, expected 20
[FAIL] navigator.credentials.create() promise is rejected if called with an aborted controller in same-origin iframe.
  promise_rejects_dom: function "function() { throw e; }" threw object "NotSupportedError: Only exactly one of 'password', 'federated', and 'publicKey' credential types are currently supported." that is not a DOMException AbortError: property "code" is equal to 9, expected 20
[FAIL] navigator.credentials.create() promise is rejected if called with an aborted signal in cross-origin iframe.
  assert_equals: expected "AbortError" but got "NotSupportedError"
[FAIL] navigator.credentials.create() promise is rejected if abort controller is aborted after call to create().
  promise_rejects_dom: function "function() { throw e; }" threw object "NetworkError: Error retrieving a token." that is not a DOMException AbortError: property "code" is equal to 19, expected 20
[FAIL] navigator.credentials.create() promise is rejected if abort controller is aborted after call to create() in cross-origin iframe.
  assert_equals: expected "AbortError" but got "NetworkError"
[FAIL] Adding mediations together with abort signal respects the abort signal.
  promise_rejects_dom: function "function() { throw e; }" threw object "NotSupportedError: Only exactly one of 'password', 'federated', and 'publicKey' credential types are currently supported." that is not a DOMException AbortError: property "code" is equal to 9, expected 20
[FAIL] Mediation is required to create a DigitalCredential.
  promise_rejects_js: function "function() { throw e; }" threw object "NetworkError: Error retrieving a token." ("NetworkError") expected instance of function "function TypeError() { [native code] }" ("TypeError")
Harness: the test ran to completion.
